Added timeout handling for reading from socketcan_native buses.
Despite the fact that the SocketscanNative_Bus.recv function has a timeout
parameter, it is ignored. It is simple to enable timeouts for a socket manually
using the settimeout() function on the socket object, but this results in an
unhandled exception in the capturePacket function. So, two exceptions are now
handled in the capturePacket function:

-BlockingIOError, which is raised when the socket has a timeout of 0.0 and you
try to read from the socket while it's empty.
-socket.timeout, which is raised when the socket has a non-zero timeout and you
try to read from the socket while it's empty.

In addition to handling these exceptions, the timeout argument to the recv
function now allows you to set the socket's timeout. I didn't like the idea of
setting the timeout every single time you call the recv function, so it only
happens when a timeout parameter is supplied. Keep in mind that
socket.settimeout(None) is still a valid function to run, as this sets the
socket to non-blocking mode. In order to do that, you'll just have to manually
execute that command before running recv().

Finally, if recv detects that a timeout occurred, it will return None, which is
consistent with the behavior of socketcan_ctypes.