Return compound error strings
It seems that the errors are sometimtes reported as a bit field and when this
happens `GetErrorText()` is unable to parse them.  If it fails reading the
error as a single value, break it down into bits and report each one on a
separate line.  For example, I just got:

```
Bus error: an error counter reached the 'heavy'/'warning' limit
Bus error: the CAN controller is in bus-off state
```