ixxat: In recv(), avoid the use of canChannelWaitRxEvent()
The function canChannelWaitRxEvent() does not always return even if there is one message in the FIFO.
This looks like a regression with at least VCI V4 4.0.482 since with VCI v3 3.5.4 there is no
problem.

This commit also refactor the recv() function in order to be more robust.

While testing this fix, I discover that canChannelReadMessage(), with VCI V4, can randomly return
the error code VCI_E_RXQUEUE_EMPTY, which by the documentation should never happen, so catch it...

Fix #201