Stop raising accuracy error for NeoVi
Will only keep logging the error that ICS lib consider critical and no
longer raise them. In most case, the error is affecting the accuracy
but the device can still be use normally (i.e. Rx buffer overflow).
In these cases, we do not want for example to kill a notifier using
this bus, etc.