Fix missing handle for CANAL API functions.
The internal reference for an opened interface via the CANAL API is
returned as an integer from the CanalOpen() DLL call in
Usb2CanAbstractionLayer.open().  It is supposed to be returned from
the abstraction layer and stored in the Usb2canBus.handle member for
later reference by all other API calls.

Add the missing return statement and adjust the docstring
accordingly.  Only valid handles (greater than 0) will be returned,
otherwise an exception is thrown already.