Fix stuct packing in build_bcm_header on 32-bit (#573)
* Fix stuct packing in build_bcm_header on 32-bit

Previously build_bcm_header relied upon adding a "0q" at the end of the
struct format string in order to force a 8-byte wide alignment on the
struct, in order to interop with the structure defined in the Linux
headers:

    struct bcm_msg_head {
      __u32 opcode;
      __u32 flags;
      __u32 count;
      struct bcm_timeval ival1, ival2;
      canid_t can_id;
      __u32 nframes;
      struct can_frame frames[0];
    };

The Python code assumed that alignof(long long) == 8 bytes in order to
accomplish this. However, on certain 32-bit platforms, this assumption
is not true, and alignof(long long) == 4.

As the struct module does not provide logic to derive alignments of its
types, this changes the packing logic to use ctypes. This exposes the
alignment of the struct members, allowing us to determine whether padding
bytes are necessary.

Fixes #470

* Call BCM factory upon module initialization

In order to prevent the BCM factory method from being called each time a
BCM header struct is created, we create the class when the module gets
created.

* Add SocketCAN interface tests for BCM factory

Add a SocketCAN interface test file with tests for validating the
BcmMsgHead class created by bcm_header_factory on various platforms.

* Clean up comments referencing packing via struct

Clean up the old comment in build_bcm_header referencing the old packing
code that called into Python's struct library. This now uses ctypes, so
and the class is created via the bcm_header_factory factory function, so
we move the comment there instead.