Enable pylint on Travis CI builds
Run pylint on all Travis CI builds for rules that we currently are
conformant to.

This adds a .pylintrc-wip file, which is intended to be a temporary
stop-gap that contains the current rules that the project is conformant
to. This allows us to incrementally enable pylint rules and have them
them enforced in subsequent code, rather than fixing the violations all
at once.

The .pylintrc file is intended to be the file containing all the
preferred set of rules that the project will eventually be compliant
against.