Add support for multiple Cyclic Messages in Task
This adds support for multiple Cyclic Messages in a Cyclic Task. The
default implementation is also changed to provide support for this, by
iterating over the list of Cyclic Messages. In order to maintain
backwards compatibility, the periodic APIs now take a CAN Message as
before, in addition to a Sequence of CAN Messages.

The SocketCAN interface takes advantage of the Linux BCM APIs to do so,
while the IXXAT interface maintains its original behaviour.

This also introduces a new example that illustrates how to use Cyclic
Messages, backed by the SocketCAN interface.

We previously tracked the can_id and arbitration_id class members due to
the ongoing deprecation of the can_id Message attribute. Now that can_id
is replaced by arbitration_id, we no longer need this in
CyclicSendTaskABC either. As such, this removes the deprecated can_id
member from the Cyclic Task.

Fixes #606