Fix iteration in Bus.stop_all_periodic_tasks
Modifying a list while iterating over it results in unspecified order. This is now fixed in `Bus.stop_all_periodic_tasks`.

What happens when an exception is thrown while stopping a task? Do we want to catch it, log it and continue? Or do we want to leave the note on unspecified behaviour in there?

Fixes #634