Add typing annotations for can.broadcastmanager
This adds typing annotations for functions in can.broadcastmanager. A
Channel definition is introduced for use in can.bus for the generic
case.

In addition, this remove the redundant typing information that was
previously in the docstring, since we now have sphinx-autodoc-typehints
to generate the types for the docs from the annotations in the function
signature.

This works towards PEP 561 compatibility.