Add typing annotations for can.message
This adds typing annotations for functions in can.message. Currently the
typing annotation for the Message data is incomplete due to upstream not
implementing the buffer protocol. Since can.Message attempts to cast the
data to a bytearray via bytearray(), the Message data can be any type
that bytearray() supports.

In addition, this remove the redundant typing information that was
previously in the docstring, since we now have sphinx-autodoc-typehints
to generate the types for the docs from the annotations in the function
signature.

This works towards PEP 561 compatibility.