Fix macOS builds
Travis CI dropped support for Python builds on macOS [0], so we just
pick a base image and use the system Python. We pin xcode9.4, which is
expected to have Python 3.7.5.

Then since our pyproject.toml file specifies a minimum Python version,
we just let poetry find it, instead of having to install a custom
Python version from Homebrew or replace the system Python.

All this is to:

  1. Get a recent Python 3
  2. Use a relatively recent coverage version (>= 5.0).
  3. Work around an inadvertently introduced ABI incompatibility
     (https://bugs.python.org/msg284585) in the Python 3.6 sqlite
     library that results in issues when running coverage tests, as
     older images use the affected Python version.

[0] https://docs.travis-ci.com/user/languages/python/