Enable GitHub Actions runners to run tests
GitHub announced Actions at GitHub Universe in 2018, however this was
somewhat limited, and support for running CI/CD pipelines was only
released in August 2019. This was in beta for a long time, and finally
has been rolled out to the general public.

This adds a minimal GitHub Actions workflow that runs all the current
tests on a build matrix consisting of {Ubuntu,macOS,Windows} and Python
3.{6,7,8}. This also seems to run much faster than the current Travis CI
pipelines, and actually lets us cover multiple versions on macOS without
a series of hacks.

Unfortunately, the kernel that GitHub Actions uses (5.0.0-1035-azure)
doesn't include the vcan kernel module, so we still need Travis CI to
run SocketCAN tests.