Fix iteration in Bus.stop_all_periodic_tasks
Fix iteration in Bus.stop_all_periodic_tasks so tasks are properly
stopped.

In addition, add a test in order to verify that Cyclic Tasks are
correctly stopped when multiple tasks are active on a particular bus.
The test is carried out via the SocketCAN interface.

Note: This is a squashed cherry pick of the following commits from
develop:
  - 8112d13ceb4493067174be899d4f014131cd4114
  - 1aa0bc64033b5fff6bc88474e7d1afb52e81ee7a

Addresses #634