Add function to list hwids of available IXXAT devices
To use multiple IXXAT devices at same time you have to know their hwids beforehand.
If it does not matter which device is used, a list of available devices is needed.

A potential use case is simulating both ends of a CAN bus with another device in between, so it does not matter which device is sender or receiver.

Usage is as easy as:
from can.interfaces.ixxat import get_ixxat_hwids
for hwid in get_ixxat_hwids():
    print("Found IXXAT with hardware id '%s'." % hwid)