Changes interface to only handle a single channel at a time
To keep with API it is best to only have a single channel handled by a canalystii.CanalystIIBus instance at a time.
There is a container to keep reference of the channels that have been created for an interface. This is done so API calls to access the interface only get done when they need to be.