removes use of the bloated pywin32 library
Most of the code I hard coded in what pywin32 does. There is a portion of the code that used wmi to enumerate usb devices and would return a piece of the instance id. I elected to use an already made library for this (pysetupdi) instead of coding it myself, which I can do if needed. This library is not listed on the python package index and is installed from the authors GitHub repository using setuptools. Collecting the instance ids using pysetupdi is going to be leaps and bounds faster then using COM objects to access the Windows WMI database. It is also going to be less prone to problems.

I have not tested the code because I do not have the devices that used pywin32. If there are people that do have these devices and are willing to report back if the code changes work or not that would be really helpful.