Add udp multicast bus (#644)
Adds a virtual bus interface udp_multicast that allows multiple processes - potentially on different hosts - to exchange CAN data. This is achieved by serializing the CAN frames using MessagePack, wrapping them in IPv6/4 packets and sending them to a multicast group.

The TTL of the IP packets is set to one by default, disallowing the packet to escape the localhost. It is configurable however, and would allow the communicating processes to run on different machines as well.