Try to find libPCBUSB.dylib before loading it
Seems like path is not always resolved when the lib is installed.
On my end the lib was installed under /usr/local/lib/. the only way to get
it imported was to find the library and give the full path to the
LoadLibrary call.

find_library need to be imported explicitely.

Signed-off-by: Stephane Dorre <stephane.dorre@gmail.com>