Adapter enumeration
Each interface should provide a means for enumerating all matching adapters. The result should be an list of adapter identifiers. An adapter identifier is not limited or equal to serial number, it can be a place in a device tree (like PCI or USB tree).

The adapter identifier then can be used with the init of the bus/interface.