Remove commented lines and update docstring
The _get_new_writer function was previously calling
`self._writer.close()`. The `self.writer` function is already being
closed in the call stack. Due to `self.writer` and `self._writer`
being linked, there was an error that the IO file was closed.

The `self._writer.close()` was commented out in a prior commit as a
proposal for the change. That comment is now removed. The docstring
is updated.