Fix race condition in back2back_test for UDP multicast bus (#1349)
* Add debug prints to back2back_test to analyze race condition

Might fail if pytest does not print stdout

* Update IPv6 back2back test to use interface-local multicast address

Under unknown conditions the IPv6 multicast frames sent during the
test are reflected back to the sender twice. This causes the test to
fail. With this commit, the multicast frames sent during the test are
constrained to the local interface to potentially avoid a double
reception.

* Revert "Add debug prints to back2back_test to analyze race condition"

This reverts commit a830c0900dee465d3d7ffa9697f49791dde722da.