Fix Bus.__new__ for PEAK CAN-FD interfaces (#1460)
* Add failing unit test to verify that issue #1485 is fixed

https://github.com/hardbyte/python-can/issues/1458

* Remove unused generic BitTiming creation in Bus.__new__

The BitTiming class is an attempt at unifying the various timing
parameters of the individual interfaces. The idea is that instead of
manually supplying multiple parameters that make up the timing
definition of the interface, one can instead supply a single instance of
the BitTiming class, which will also automatically calculate derivative
values from its input.

At the moment, this class is only used by two interfaces: CANtact and
CANanalystii. Both either accept a single bitrate or a BitTiming
instance. The latter will overrule the former.

The code that is removed with this commit is part of the generic
Bus.__new__ constructor. The removed code searches the set of kwargs
parameters for timing-related values. If it finds at least one such
value, it creates a BitTiming class instance and adds it to the list of
parameters. However, it breaks compatibility with the PEAK interface,
see issue #1458. Additionally, the code in question is generic and
applies to all interfaces. Instantiating a class here is prone to issues
since it must be generic enough to fit all use cases. A better approach
would be to simply forward the parameters as was done previously and
leave it up to the individual interfaces to handle things properly.

* Format code with black

Co-authored-by: lumagi <lumagi@users.noreply.github.com>