PCAN: Fix detection of lib on Windows on ARM (#1463)
Before this commit detection of the PCAN DLL would fail on Windows on
ARM, regardless if you used the native Python version, or the x64
version of Python. After this fix detection should work properly as long
as the PCAN library for your version of Python is listed first in the
`PATH` variable. The default:

1. `C:\Program Files\PEAK-System\PEAK-Drivers 4\APIs\ARM64\` before
2. `C:\Program Files\PEAK-System\PEAK-Drivers 4\APIs\x64\`

should work if you use the (native) ARM version of Python. If you
reorder these paths, then loading the library works in the x64 version
of Python.

This commit closes #1461.