udp_multicast interface: support windows (#1914)
* support windows

Neither the SO_TIMESTAMPNS / via recvmsg() method, nor the SIOCGSTAMP / ioctl() method for obtaining the packet timestamp is supported on Windows.  This code adds a fallback to datetime, and switches to recvfrom() so that the udp_multicast bus becomes usable on windows.

* clean-up example

on windows, the example would otherwise cause an _enter_buffered_busy fatal error, notifier shutdown and bus shutdown are racing eachother...

* fix double inversion

* remove unused import

* datetime to time.time()

* add msgpack dependency for windows

* enable udp_multicast back2back tests on windows

* handle WSAEINVAL like EINVAL

* avoid potential AttributeError on Linux

* simplify unittest skip condition

* fix formatting issue